feat: cjsToEsm.cjs - rewrite it to it's equivalent in ECMAScript notation
cjsToEsm.cjs - rewrite it to it's equivalent in ECMAScript notation (and rename it to esm.mjs)
